requirements.txt
numpy>=1.26
Pillow>=10.0
scipy>=1.11
scikit-image>=0.22